server/requirements.txt
fastapi
mcp
openai
pandas
pydantic
python-dotenv
uvicorn
